Changed the way amounts are stored. Everything is now stored in pence and have updated the number formatter to match.

+++ database/migrations/2017_07_14_132128_create_invoices_table.php
@@ -21,10 +21,10 @@ class CreateInvoicesTable extends Migration
             $table->integer('invoice_recurring_id')->unsigned()->default(0);
             $table->string('number');
             $table->text('recipient')->nullable();
-            $table->decimal('net', 13, 3)->default(0);
-            $table->decimal('tax', 13, 3)->default(0);
-            $table->decimal('total', 13, 3)->default(0);
-            $table->decimal('balance', 13, 3)->default(0);
+            $table->integer('net')->default(0);
+            $table->integer('tax')->default(0);
+            $table->integer('total')->default(0);
+            $table->integer('balance')->default(0);
             $table->text('terms')->nullable();
             $table->string('key');
             $table->timestamp('due_at')->nullable();
