+++ database/migrations/2017_07_21_080740_create_tenancies_table.php
@@ -18,7 +18,7 @@ class CreateTenanciesTable extends Migration
             $table->integer('user_id')->unsigned();
             $table->integer('property_id')->unsigned();
             $table->integer('service_id')->unsigned();
-            $table->decimal('rent_balance', 12, 3)->default(0);
+            $table->integer('rent_balance')->default(0);
             $table->integer('is_overdue')->default(0);
             $table->date('started_on')->nullable();
             $table->date('vacated_on')->nullable();
